--- services/service_log_improvement/test/modified_file.txt
@@ -1,26 +1,0 @@
-public class Main {
-
-    public static void main(String[] args) {
-        // This is the entry point of the application
-        System.out.println("Welcome to the Java application!");
-
-        // Example of how to use the application logic
-        if (args.length > 0) {
-            System.out.println("This is my new log ");
-            for (String arg : args) {
-                System.out.println(arg);
-            }
-        } else {
-            System.out.println("No arguments received.");
-        }
-
-        // Example method call to demonstrate further functionality
-        int result = addNumbers(5, 10);
-        System.out.println("Sum of 5 and 10 is: " + result);
-    }
-
-    // Example utility method
-    public static int addNumbers(int a, int b) {
-        return a + b;
-    }
-}
