--- services/service_log_improvement/test/test2/diff2.txt
@@ -1,9 +1,0 @@
-diff --git a/MainApplication.java b/MainApplication.java
-index 9a01aa5..11801f2 100644
---- a/MainApplication.java
-+++ b/MainApplication.java
-@@ -9 +9 @@ public class Main {
--            System.out.println("Arguments recei: ");
-+            System.out.println("Arguments receie: ");
-@@ -14,0 +15 @@ public class Main {
-+            System.out.println("No arguments receivedddddd.");
